+++ docs-gen/utils/generate-docs.sh
@@ -93,6 +93,8 @@ cat "$README_CHAPTERS_DIR"/mbx-intro.md >>README_TEMP.md
 
 cat "$README_CHAPTERS_DIR"/building-process-heading.md >>README_TEMP.md
 cat "$README_CHAPTERS_DIR"/building-process-shared-props.md >>README_TEMP.md
+echo -n "\n\n<br>\n\n" >>README_TEMP.md
+cat "$README_CHAPTERS_DIR"/building-process-accessibility-props.md >>README_TEMP.md
 
 echo -n "\n\n<br>\n\n" >>README_TEMP.md