+++ src/components/molecules/FormField/component.tsx
@@ -59,7 +59,7 @@ const FormFieldInternalComponent: MbxUiReactiveComponent<
         {...headerProps}
       >
         {header}
-      </Container>
+      </Container>,
     );
   }
 
@@ -107,7 +107,7 @@ const FormFieldInternalComponent: MbxUiReactiveComponent<
 
         setValue(formattedValue);
       }}
-    />
+    />,
   );
 
   components.push(
@@ -122,7 +122,7 @@ const FormFieldInternalComponent: MbxUiReactiveComponent<
       key="form_field_error_box"
     >
       {errorLabel}
-    </Container>
+    </Container>,
   );
 
   return components;